pyproject.toml
[tool.poetry]
name = "chatglm-6b-api"
version = "0.1.0"
description = ""
authors = ["XingKaiXin <xingkaixin@gmail.com>"]
readme = "README.md"
packages = [{include = "chatglm_6b_api"}]

[tool.poetry.dependencies]
python = "^3.8"
transformers = "4.27.1"
protobuf = "^4.22.3"
cpm-kernels = "^1.0.11"
sentencepiece = "^0.1.98"
text2vec = "^1.1.8"
pydantic = "^1.10.7"
pyyaml = "^6.0"
fastapi = "^0.95.1"
sse-starlette = "^1.3.4"



[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.flake8]
max-line-length = 120
exclude = "**/__init__.py"

[tool.pylint.messages_control]
disable = "invalid-name, redefined-outer-name, missing-module-docstring, missing-function-docstring, missing-class-docstring, W0703"
logging-format-style = "percent"

[tool.pylint.design]
min-public-methods = 0

[tool.pylint.format]
max-line-length = 120
indent-string = '    '
single-quotes = true

[tool.pylint.extensions]
extension-pkg-allow-list = ["pydantic"]

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88